ー連の流れ

+++ JPHACKS2018-NG1806/Assets/Temp.cs
@@ -7,7 +7,7 @@ public class Temp : MonoBehaviour {
 
 
     public static string sendname;
-    public static int sendnumber;
+    public static long sendnumber;
     public static long nowobjnum;
     public static string look_name;
     public static string look_obj;
